+++ rust/operator-binary/src/controller.rs
@@ -13,9 +13,11 @@ use stackable_nifi_crd::{
     HTTPS_PORT_NAME, METRICS_PORT, METRICS_PORT_NAME, PROTOCOL_PORT, PROTOCOL_PORT_NAME,
 };
 use stackable_nifi_crd::{APP_NAME, BALANCE_PORT, BALANCE_PORT_NAME};
+use stackable_operator::cluster_resources::ClusterResources;
 use stackable_operator::commons::resources::{NoRuntimeLimits, Resources};
 use stackable_operator::config::merge::Merge;
 use stackable_operator::k8s_openapi::api::apps::v1::StatefulSetUpdateStrategy;
+use stackable_operator::kube::Resource;
 use stackable_operator::role_utils::Role;
 use stackable_operator::{
     builder::{ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder},
@@ -49,6 +51,7 @@ use std::{
 use strum::{EnumDiscriminants, IntoStaticStr};
 use tracing::Instrument;
 
+const CONTROLLER_NAME: &str = "nifi-operator";
 const FIELD_MANAGER_SCOPE: &str = "nificluster";
 const STACKABLE_TOOLS_IMAGE: &str = "docker.stackable.tech/stackable/tools:0.2.0-stackable0";
 
@@ -73,6 +76,14 @@ pub enum Error {
     ObjectHasNoNamespace,
     #[snafu(display("failed to calculate global service name"))]
     GlobalServiceNameNotFound,
+    #[snafu(display("failed to create cluster resources"))]
+    CreateClusterResources {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to delete orphaned resources"))]
+    FinalizeClusterResources {
+        source: stackable_operator::error::Error,
+    },
     #[snafu(display("failed to apply global Service"))]
     ApplyRoleService {
         source: stackable_operator::error::Error,
@@ -134,11 +145,6 @@ pub enum Error {
         obj_ref: ObjectRef<NifiCluster>,
         selector: LabelSelector,
     },
-    #[snafu(display("Failed to find service {obj_ref}"))]
-    MissingService {
-        source: stackable_operator::error::Error,
-        obj_ref: ObjectRef<Service>,
-    },
     #[snafu(display("Failed to materialize authentication config element from k8s"))]
     MaterializeAuthConfig {
         source: stackable_nifi_crd::authentication::Error,
@@ -167,11 +173,6 @@ pub async fn reconcile_nifi(nifi: Arc<NifiCluster>, ctx: Arc<Ctx>) -> Result<Act
     let nifi_product_version = nifi
         .product_version()
         .context(NifiVersionParseFailureSnafu)?;
-    let namespace = &nifi
-        .metadata
-        .namespace
-        .clone()
-        .with_context(|| ObjectHasNoNamespaceSnafu {})?;
 
     tracing::info!("Checking for sensitive key configuration");
     check_or_generate_sensitive_key(client, &nifi).await?;
@@ -189,18 +190,15 @@ pub async fn reconcile_nifi(nifi: Arc<NifiCluster>, ctx: Arc<Ctx>) -> Result<Act
         .map(Cow::Borrowed)
         .unwrap_or_default();
 
+    let mut cluster_resources =
+        ClusterResources::new(APP_NAME, CONTROLLER_NAME, &nifi.object_ref(&()))
+            .context(CreateClusterResourcesSnafu)?;
+
     let node_role_service = build_node_role_service(&nifi)?;
-    client
-        .apply_patch(FIELD_MANAGER_SCOPE, &node_role_service, &node_role_service)
+    let updated_role_service = cluster_resources
+        .add(client, &node_role_service)
         .await
         .context(ApplyRoleServiceSnafu)?;
-
-    let updated_role_service = client
-        .get(&nifi.name(), nifi.namespace().as_deref())
-        .await
-        .with_context(|_| MissingServiceSnafu {
-            obj_ref: ObjectRef::new(&nifi.name()).within(namespace),
-        })?;
 
     for (rolegroup_name, rolegroup_config) in nifi_node_config.iter() {
         let rg_span = tracing::info_span!("rolegroup_span", rolegroup = rolegroup_name.as_str());
@@ -244,27 +242,27 @@ pub async fn reconcile_nifi(nifi: Arc<NifiCluster>, ctx: Arc<Ctx>) -> Result<Act
 
             let reporting_task_job = build_reporting_task_job(&nifi, &rolegroup)?;
 
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_service, &rg_service)
+            cluster_resources
+                .add(client, &rg_service)
                 .await
                 .with_context(|_| ApplyRoleGroupServiceSnafu {
                     rolegroup: rolegroup.clone(),
                 })?;
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_configmap, &rg_configmap)
+            cluster_resources
+                .add(client, &rg_configmap)
                 .await
                 .with_context(|_| ApplyRoleGroupConfigSnafu {
                     rolegroup: rolegroup.clone(),
                 })?;
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_log_configmap, &rg_log_configmap)
+            cluster_resources
+                .add(client, &rg_log_configmap)
                 .await
                 .with_context(|_| ApplyRoleGroupConfigSnafu {
                     rolegroup: rolegroup.clone(),
                 })?;
 
-            client
-                .apply_patch(FIELD_MANAGER_SCOPE, &rg_statefulset, &rg_statefulset)
+            cluster_resources
+                .add(client, &rg_statefulset)
                 .await
                 .with_context(|_| ApplyRoleGroupStatefulSetSnafu {
                     rolegroup: rolegroup.clone(),
@@ -283,6 +281,11 @@ pub async fn reconcile_nifi(nifi: Arc<NifiCluster>, ctx: Arc<Ctx>) -> Result<Act
         .instrument(rg_span)
         .await?
     }
+
+    cluster_resources
+        .finalize(client)
+        .await
+        .context(FinalizeClusterResourcesSnafu)?;
 
     Ok(Action::await_change())
 }
@@ -305,6 +308,7 @@ pub fn build_node_role_service(nifi: &NifiCluster) -> Result<Service> {
                 nifi,
                 APP_NAME,
                 nifi.image_version().context(NifiVersionParseFailureSnafu)?,
+                CONTROLLER_NAME,
                 &role_name,
                 "global",
             )
@@ -353,6 +357,7 @@ fn build_node_rolegroup_log_config_map(
                     nifi,
                     APP_NAME,
                     nifi.image_version().context(NifiVersionParseFailureSnafu)?,
+                    CONTROLLER_NAME,
                     &rolegroup.role,
                     &rolegroup.role_group,
                 )
@@ -395,6 +400,7 @@ async fn build_node_rolegroup_config_map(
                     nifi,
                     APP_NAME,
                     nifi.image_version().context(NifiVersionParseFailureSnafu)?,
+                    CONTROLLER_NAME,
                     &rolegroup.role,
                     &rolegroup.role_group,
                 )
@@ -461,6 +467,7 @@ fn build_node_rolegroup_service(
                 nifi,
                 APP_NAME,
                 nifi.image_version().context(NifiVersionParseFailureSnafu)?,
+                CONTROLLER_NAME,
                 &rolegroup.role,
                 &rolegroup.role_group,
             )
@@ -759,6 +766,7 @@ fn build_node_rolegroup_statefulset(
                 nifi,
                 APP_NAME,
                 image_version,
+                CONTROLLER_NAME,
                 &rolegroup_ref.role,
                 &rolegroup_ref.role_group,
             )
@@ -856,6 +864,7 @@ fn build_node_rolegroup_statefulset(
                 nifi,
                 APP_NAME,
                 image_version,
+                CONTROLLER_NAME,
                 &rolegroup_ref.role,
                 &rolegroup_ref.role_group,
             )